// src/app/enums/enum.ts
export enum FileType {
  PRIVATE = 'private',
  PUBLIC = 'public',
}

export enum FileResizeFit {
  COVER = 'cover',
  CONTAIN = 'contain',
  FILL = 'fill',
  INSidE = 'inside',
  OUTSidE = 'outside',
}

export enum FileResizePosition {
  CENTER = 'center',
  TOP = 'top',
  RIGHT = 'right',
  BOTTOM = 'bottom',
  LEFT = 'left',
  RIGHT_TOP = 'right top',
  RIGHT_BOTTOM = 'right bottom',
  LEFT_TOP = 'left top',
  LEFT_BOTTOM = 'left bottom',
  ENTROPY = 'entropy',
  ATTENTION = 'attention',
}